fix: Work around libqdx encoding issues
Use the new JsonOptions struct to apply special processing to certain
enums that are encoded as ints, and which have one or more missing or
incorrect discriminants in their ouroboros type. The latter can occur if
an enum has a discriminant outside the range 0-255, since the
derive(TypeInfo) macro only recognises those which fit into a u8.

Also convert ouroboros Alias types to just their inner type when going
from ouroboros -> rex. This fixes an issue with the use of constructors
in tengu.